Allow fractional shares (#445)
* Allow fractional shares

* Allow fractional shares

+++ src/portfolios/domain/dto/test/upsert-position-dto.factory.ts
@@ -9,8 +9,8 @@ export function upsertPositionDtoFactory(
 ): UpsertPositionDto {
   return <UpsertPositionDto>{
     symbol: symbol ?? faker.word.words(),
-    targetWeight: targetWeight ?? faker.number.int({ min: 5, max: 10 }),
-    shares: shares ?? faker.number.int({ min: 10, max: 100 }),
+    targetWeight: targetWeight ?? faker.number.float({ min: 5, max: 10 }),
+    shares: shares ?? faker.number.float({ min: 10, max: 100 }),
     blocked: blocked ?? faker.datatype.boolean(),
   };
 }
